Modify Function NFC, Firebase, pantalla de recarga, modify model operation

+++ app/src/main/java/com/puyodev/luka/model/User.kt
@@ -5,5 +5,5 @@ import com.google.firebase.firestore.DocumentId
 data class User(
     @DocumentId val id: String = "",
     val username: String = "",
-    val lukitas: Double = 0.00
+    val lukitas: Int = 0
 )
